view Cart item working. bug at the add item on first item on combo

+++ app/src/main/java/za/grounded/jwc_app/database/daos/TransactionDao.java
@@ -18,7 +18,7 @@ public interface TransactionDao {
     @Query("UPDATE transactions " +
             "SET total = :total " +
             "WHERE room_id = :id")
-    void updateTransactionTotal(Long id, Long total);
+    void updateTransactionTotal(Long id, Double total);
 
     @Query("UPDATE transactions " +
             "SET transactionDate = :transactionDate " +
